pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aslookup"
version = "2.0.0"
authors = [
    { name="Darren Spruell", email="phatbuckett@gmail.com" },
]
description = "IP to ASN routing data query script"
readme = "README.md"
dependencies = [
    "aiodns",
    "defang",
    "dnspython",
    "pytricia",
]
classifiers = [
    # "Development Status :: 3 - Alpha",
    # "Development Status :: 4 - Beta",
    "Development Status :: 5 - Production/Stable",
    # "Development Status :: 6 - Mature",
    "License :: OSI Approved :: ISC License (ISCL)",
    "Programming Language :: Python :: 3",
]

[project.scripts]
as-lookup = "aslookup.cli:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "pytest-cov",
]

[project.urls]
"Bug Tracker" = "https://github.com/dspruell/aslookup/issues"
"Homepage" = "https://github.com/dspruell/aslookup"
